script/requirements.txt
aiofiles==24.1.0
anyio==4.9.0
async-property==0.2.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
deprecation==2.1.0
Faker==37.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jwcrypto==1.5.6
packaging==25.0
pycparser==2.22
python-keycloak==5.5.1
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
